Improve users API
