+++ app/views/api/users/activations/create.jbuilder
