+++ app/views/api/users/show.jbuilder
