+++ spec/requests/api/users/authorizations_request_spec.rb
@@ -1,0 +1,66 @@
+require 'rails_helper'
+require 'shared_examples_for_failures'
+
+RSpec.describe Api::Users::AuthorizationsController, type: :request do
+
+  describe 'POST #create' do
+    context 'with valid username and password' do
+      let(:payload) { { username: 'john', password: 'secret' } }
+
+      before do
+        Timecop.freeze Date.new(2017)
+        create :user, :activated, username: 'john', password: 'secret'
+        post '/api/users/authorizations', params: payload
+      end
+
+      after do
+        Timecop.return
+      end
+
+      it 'succeeds' do
+        expect(response).to have_http_status(:ok)
+      end
+
+      it 'matches the users/authorizations/create format' do
+        expect(response).to match_response_schema('users/authorizations/create')
+      end
+
+      it 'returns a token valid for 1 month' do
+        token = JSON.parse(response.body)['token']
+        decoded_token = JsonWebToken.decode(token)
+        expect(decoded_token[:exp]).to eq(1.month.from_now.to_i)
+      end
+    end
+
+    context 'with inactive user' do
+      let(:payload) { { username: 'john', password: 'secret' } }
+
+      before do
+        create :user, username: 'john', password: 'secret'
+        post '/api/users/authorizations', params: payload
+      end
+
+      it_behaves_like 'failures', :unauthorized, 'custom_error', {
+        message: 'Bad credentials',
+        code: 'login_failed',
+        resource: 'User',
+      }
+    end
+
+    context 'with invalid password' do
+      let(:payload) { { username: 'john', password: 'wrong secret' } }
+
+      before do
+        create :user, :activated, username: 'john', password: 'secret'
+        post '/api/users/authorizations', params: payload
+      end
+
+      it_behaves_like 'failures', :unauthorized, 'custom_error', {
+        message: 'Bad credentials',
+        code: 'login_failed',
+        resource: 'User',
+      }
+    end
+  end
+
+end
