+++ spec/requests/api/users_request_spec.rb
@@ -61,167 +61,6 @@ RSpec.describe Api::UsersController, type: :request do
     end
   end
 
-  describe 'PATCH #activate' do
-    let(:user) { create :user, email: 'john@doe.com' }
-
-    context 'with valid attributes' do
-      before do
-        Timecop.freeze Date.new(2017)
-        payload = {
-          username: 'john',
-          password: 'secret',
-        }
-        post "/api/users/#{ user.activation_token }/activate", params: { user: payload }
-      end
-
-      after do
-        Timecop.return
-      end
-
-      it 'succeeds' do
-        expect(response).to have_http_status(:ok)
-      end
-
-      it 'matches the users/activate schema' do
-        expect(response).to match_response_schema('users/activate')
-      end
-
-      it 'activates the user' do
-        expect(user.reload.activation_state).to eq('active')
-      end
-
-      it 'returns the new user' do
-        contact = JSON.parse(response.body)['user']
-        expect(contact['id']).not_to be_nil
-        expect(contact['username']).to eq('john')
-        expect(contact['email']).to eq('john@doe.com')
-      end
-
-      it 'returns a token valid for 1 month' do
-        token = JSON.parse(response.body)['token']
-        decoded_token = JsonWebToken.decode(token)
-        expect(decoded_token[:exp]).to eq(1.month.from_now.to_i)
-      end
-    end
-
-    context 'with missing attributes' do
-      before do
-        payload = {
-          password: 'secret',
-        }
-        post "/api/users/#{ user.activation_token }/activate", params: { user: payload }
-      end
-
-      it_behaves_like 'missing param failures', 'User', 'username'
-    end
-
-    context 'with invalid username' do
-      before do
-        payload = {
-          username: 'John Doe',
-          password: 'secret',
-        }
-        post "/api/users/#{ user.activation_token }/activate", params: { user: payload }
-      end
-
-      it_behaves_like 'validation failed failures', 'User', { username: ['invalid'] }
-    end
-
-    context 'with too long username' do
-      before do
-        payload = {
-          username: 'johnjohnjohnjohnjohnjohnjohn',
-          password: 'secret',
-        }
-        post "/api/users/#{ user.activation_token }/activate", params: { user: payload }
-      end
-
-      it_behaves_like 'validation failed failures', 'User', { username: ['too_long'] }
-    end
-
-    context 'with existing username' do
-      before do
-        create :user, username: 'john'
-        payload = {
-          username: 'john',
-          password: 'secret',
-        }
-        post "/api/users/#{ user.activation_token }/activate", params: { user: payload }
-      end
-
-      it_behaves_like 'validation failed failures', 'User', { username: ['taken'] }
-    end
-
-    context 'with invalid token' do
-      before do
-        payload = { password: 'secret' }
-        post "/api/users/not_the_token/activate", params: { user: payload }
-      end
-
-      it_behaves_like 'not found failures', 'User'
-    end
-  end
-
-  describe 'POST #authorize' do
-    context 'with valid username and password' do
-      let(:payload) { { username: 'john', password: 'secret' } }
-
-      before do
-        Timecop.freeze Date.new(2017)
-        create :user, :activated, username: 'john', password: 'secret'
-        post '/api/users/authorize', params: payload
-      end
-
-      after do
-        Timecop.return
-      end
-
-      it 'succeeds' do
-        expect(response).to have_http_status(:ok)
-      end
-
-      it 'matches the users/authorize format' do
-        expect(response).to match_response_schema('users/authorize')
-      end
-
-      it 'returns a token valid for 1 month' do
-        token = JSON.parse(response.body)['token']
-        decoded_token = JsonWebToken.decode(token)
-        expect(decoded_token[:exp]).to eq(1.month.from_now.to_i)
-      end
-    end
-
-    context 'with inactive user' do
-      let(:payload) { { username: 'john', password: 'secret' } }
-
-      before do
-        create :user, username: 'john', password: 'secret'
-        post '/api/users/authorize', params: payload
-      end
-
-      it_behaves_like 'failures', :unauthorized, 'custom_error', {
-        message: 'Bad credentials',
-        code: 'login_failed',
-        resource: 'User',
-      }
-    end
-
-    context 'with invalid password' do
-      let(:payload) { { username: 'john', password: 'wrong secret' } }
-
-      before do
-        create :user, :activated, username: 'john', password: 'secret'
-        post '/api/users/authorize', params: payload
-      end
-
-      it_behaves_like 'failures', :unauthorized, 'custom_error', {
-        message: 'Bad credentials',
-        code: 'login_failed',
-        resource: 'User',
-      }
-    end
-  end
-
   describe 'GET #me' do
     let(:user) { create :user }
     let(:token) { user.token }
