+++ spec/support/api/schemas/users/activations/create.json
